View (description) file for farmer and user are done and some design changes

+++ app/Http/Controllers/Admin/UserController.php
@@ -21,8 +21,13 @@ class UserController extends Controller
      */
     public function index()
     {
-        $users = User::latest()->get();
-        return view('admin.user.index', compact('users'));
+        /* show all users */
+        if(auth()->user()->can('view_user')){
+            $users = User::latest()->get();
+            return view('admin.user.index', compact('users'));
+        }
+        abort(403);
+        
     }
 
     /**
@@ -32,9 +37,14 @@ class UserController extends Controller
      */
     public function create()
     {
-        $branches = Branch::all();
-        $roles = Role::all();
-        return view('admin.user.create', compact('branches', 'roles'));
+        /* CREATE   USER */
+        if(auth()->user()->can('create_user')){
+            $branches = Branch::all();
+            $roles = Role::all();
+            return view('admin.user.create', compact('branches', 'roles'));
+        }
+        abort(403);
+        
     }
 
     /**
@@ -45,29 +55,34 @@ class UserController extends Controller
      */
     public function store(UserStoreRequest $request)
     {
-        /* Insert User */
-        $user = User::create([
-            'branch_id'        =>      $request->branch,
-            'name'             =>      $request->user,
-            'username'         =>      $request->username,
-            'phone1'           =>      $request->phone1,
-            'phone2'           =>      $request->phone2,
-            'email'            =>      $request->email,
-            'password'         =>      bcrypt($request->password),
-            'address'          =>      $request->address,
-            'status'           =>      'active',
-        ]);
+        /* INSERT USER */
+        if(auth()->user()->can('create_user')){
+            $user = User::create([
+                'branch_id'        =>      $request->branch,
+                'name'             =>      $request->user,
+                'username'         =>      $request->username,
+                'phone1'           =>      $request->phone1,
+                'phone2'           =>      $request->phone2,
+                'email'            =>      $request->email,
+                'password'         =>      bcrypt($request->password),
+                'address'          =>      $request->address,
+                'status'           =>      'active',
+            ]);
 
-        /* assigning roles to the user */
-        if($user){
-            $user->roles()->attach($request->roles);
-        }
+            /* assigning roles to the user */
+            if($user){
+                $user->roles()->attach($request->roles);
+            }
 
-        /* Check famer insertion  and Toastr */
-        if($user){
-            Toastr::success('User Inserted Successfully', 'Success');
-            return redirect()->route('admin.user.index');
+            /* Check famer insertion  and Toastr */
+            if($user){
+                Toastr::success('User Inserted Successfully', 'Success');
+                return redirect()->route('admin.user.index');
+            }
         }
+        abort(403);
+
+        
     }
 
     /**
@@ -78,7 +93,8 @@ class UserController extends Controller
      */
     public function show($id)
     {
-        //
+        $user = User::findOrFail($id);
+        return view('admin.user.view', compact('user'));
     }
 
     /**
@@ -89,10 +105,13 @@ class UserController extends Controller
      */
     public function edit($id)
     {
-        $roles = Role::all();
-        $branches = Branch::all();
-        $user = User::findOrFail($id);
-        return view('admin.user.edit', compact('user','branches', 'roles'));
+        if(auth()->user()->can('edit_user')){
+            $roles = Role::all();
+            $branches = Branch::all();
+            $user = User::findOrFail($id);
+            return view('admin.user.edit', compact('user','branches', 'roles'));
+        }
+        abort(403);
     }
 
     /**
@@ -104,45 +123,49 @@ class UserController extends Controller
      */
     public function update(UserUpdateRequest $request, $id)
     {
-        $user = User::findOrFail($id);
-        /* update user with password */
-        if(!empty($request->password)){
-            $user->update([
-                'branch_id'        =>      $request->branch,
-                'name'             =>      $request->user,
-                'username'         =>      $request->username,
-                'phone1'           =>      $request->phone1,
-                'phone2'           =>      $request->phone2,
-                'email'            =>      $request->email,
-                'password'         =>      bcrypt($request->password),
-                'address'          =>      $request->address,
-                'status'           =>      $request->status,
-            ]);
-        } 
-        /* update user without password */
-        else{
-            $user->update([
-                'branch_id'        =>      $request->branch,
-                'name'             =>      $request->user,
-                'username'         =>      $request->username,
-                'phone1'           =>      $request->phone1,
-                'phone2'           =>      $request->phone2,
-                'email'            =>      $request->email,
-                'address'          =>      $request->address,
-                'status'           =>      $request->status,
-            ]);
-        }
+        /* UPDATE USER */
+        if(auth()->user()->can('edit_user')){
+            $user = User::findOrFail($id);
+            /* update user with password */
+            if(!empty($request->password)){
+                $user->update([
+                    'branch_id'        =>      $request->branch,
+                    'name'             =>      $request->user,
+                    'username'         =>      $request->username,
+                    'phone1'           =>      $request->phone1,
+                    'phone2'           =>      $request->phone2,
+                    'email'            =>      $request->email,
+                    'password'         =>      bcrypt($request->password),
+                    'address'          =>      $request->address,
+                    'status'           =>      $request->status,
+                ]);
+            } 
+            /* update user without password */
+            else{
+                $user->update([
+                    'branch_id'        =>      $request->branch,
+                    'name'             =>      $request->user,
+                    'username'         =>      $request->username,
+                    'phone1'           =>      $request->phone1,
+                    'phone2'           =>      $request->phone2,
+                    'email'            =>      $request->email,
+                    'address'          =>      $request->address,
+                    'status'           =>      $request->status,
+                ]);
+            }
 
-        /* update roles to the user */
-        if($user){
-            $user->roles()->sync($request->roles);
-        }
+            /* update roles to the user */
+            if($user){
+                $user->roles()->sync($request->roles);
+            }
 
-        /* Check famer insertion  and Toastr */
-        if($user){
-            Toastr::success('User Updated Successfully', 'Success');
-            return redirect()->route('admin.user.index');
+            /* Check famer insertion  and Toastr */
+            if($user){
+                Toastr::success('User Updated Successfully', 'Success');
+                return redirect()->route('admin.user.index');
+            }
         }
+        abort(403);
     }
 
     /**
@@ -153,11 +176,14 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
-        $userDelete = User::findOrFail($id)->delete();
+        if(auth()->user()->can('delete_user')){
+            $userDelete = User::findOrFail($id)->delete();
         
-        if($userDelete){
-            Toastr::success('User Deleted Successfully', 'Success');
-            return redirect()->route('admin.user.index');
+            if($userDelete){
+                Toastr::success('User Deleted Successfully', 'Success');
+                return redirect()->route('admin.user.index');
+            }
         }
+        abort(403);
     }
 }
